+++ src/DevTools/CreateDBClasses.php
@@ -95,8 +95,8 @@ class CreateDBClasses extends ConnectMySQL {
 
     public function prepare_table_php_class_creation( bool $return = false ): bool|array {
         if ( !defined( 'TBL_PFX' ) ) {
-            define( 'TBL_PFX', '' ); // Hide the error.
-            throw new Exception( "Table Prefex TBL_PFX not defined" );
+            define( 'TBL_PFX', $_ENV['TABLE_PREFIX'] ?? '' ); // Hide the error.
+            // throw new Exception( "Table Prefex TBL_PFX not defined" );
         }
         /**
          * String cleaning of the table name as required.